chore: added BeleuchtungLayerInfo component, restructured BeleuchtungInfo Components
