+++ src/lang/de.json
@@ -231,7 +231,8 @@
   "ch.sbb.nationalpark": "Schweizer Nationalpark",
   "ch.sbb.bln": "Landschaften- und Naturdenkmäler",
   "Bahnhofklasse": "Bahnhofklasse",
-  "Bahnhof-Beleuchtungsklassen": "Bahnhof Beleuchtungsklassen",
+  "Bahnhof Beleuchtungsklassen": "Bahnhof Beleuchtungsklassen",
+  "Bahnhof Beleuchtungsklasse": "Bahnhof Beleuchtungsklasse",
   "ch.sbb.beleuchtungsstaerken.info": "Werkzeug für Grundlagengespräche u.a. mit Behörden, Gemeinden, Projekt- und Anlagenverantwortlichen über die Spezifikationen von Beleuchtungsanlagen.",
   "Beleuchtungsstärke": "Beleuchtungsstärke",
   "Hohes Personenaufkommen": "Hohes Personenaufkommen",
