+++ src/lang/en.json
@@ -236,12 +236,13 @@
   "ch.sbb.nationalpark": "Schweizer Nationalpark",
   "ch.sbb.bln": "Landschaften- und Naturdenkmäler",
   "Bahnhofklasse": "Station class",
-  "Bahnhof-Beleuchtungsklassen": "Station lighting classes",
+  "Bahnhof Beleuchtungsklasse": "Station lighting class",
+  "Bahnhof Beleuchtungsklassen": "Station lighting classes",
   "Beleuchtungsstärke": "Lighting intensity",
-  "Hohes Personenaufkommen": "High Traffic",
-  "Mittleres Personenaufkommen": "Medium Traffic",
-  "Geringes Personenaufkommen": "Little Traffic",
-  "Sehr geringes Personenaufkommen": "Very little Traffic",
+  "Hohes Personenaufkommen": "High volume of people",
+  "Mittleres Personenaufkommen": "Average number of people",
+  "Geringes Personenaufkommen": "Low volume of people",
+  "Sehr geringes Personenaufkommen": "Very low volume of people",
   "Passagiere/Tag": "Passengers / day",
   "Damit sie ihre Zeichnung erneut bearbeiten können, müssen sie diesen Link speichern.": "You can save this link to edit your drawing again.",
   "departure_url": "http://fahrplan.sbb.ch/bin/stboard.exe/en?boardType=dep&editStation=yes&input=Lausanne&start=yes",
