+++ src/lang/fr.json
@@ -252,12 +252,13 @@
   "ch.sbb.nationalpark": "Schweizer Nationalpark",
   "ch.sbb.bln": "Landschaften- und Naturdenkmäler",
   "Bahnhofklasse": "Classe de gare",
-  "Bahnhof-Beleuchtungsklassen": "Classes d'éclairage des gares",
+  "Bahnhof Beleuchtungsklassen": "Classes d'éclairage des gares",
+  "Bahnhof Beleuchtungsklasse": "Classe d'éclairage de la gare",
   "Beleuchtungsstärke": "Intensité d'éclairage",
-  "Hohes Personenaufkommen": "Trafic élevé",
-  "Mittleres Personenaufkommen": "Trafic moyen",
-  "Geringes Personenaufkommen": "Peu de trafic",
-  "Sehr geringes Personenaufkommen": "Très peu de trafic",
+  "Hohes Personenaufkommen": "Nombre important de passagers",
+  "Mittleres Personenaufkommen": "Nombre moyen de passagers",
+  "Geringes Personenaufkommen": "Faible nombre de passagers",
+  "Sehr geringes Personenaufkommen": "Très faible nombre de passagers",
   "Passagiere/Tag": "Passagers / jour",
   "Damit sie ihre Zeichnung erneut bearbeiten können, müssen sie diesen Link speichern.": "Pour rééditer votre dessin, vous pouvez enregistrer ce lien.",
   "Das Thema existiert nicht oder Sie haben nicht die notwendigen Rechte": "Le thème n’existe pas ou vous ne disposez pas des autorisations requises",
