+++ src/lang/it.json
@@ -253,12 +253,13 @@
   "ch.sbb.nationalpark": "Schweizer Nationalpark",
   "ch.sbb.bln": "Landschaften- und Naturdenkmäler",
   "Bahnhofklasse": "Classe stazione",
-  "Bahnhof-Beleuchtungsklassen": "Classe di illuminazione delle stazioni ferroviarie",
+  "Bahnhof Beleuchtungsklassen": "Classi di illuminazione delle stazioni ferroviarie",
+  "Bahnhof Beleuchtungsklasse": "Classe di illuminazione stazione",
   "Beleuchtungsstärke": "Intensitá illuminazione",
-  "Hohes Personenaufkommen": "Traffico elevato",
-  "Mittleres Personenaufkommen": "Traffico medio",
-  "Geringes Personenaufkommen": "Poco traffico",
-  "Sehr geringes Personenaufkommen": "Molto poco traffico",
+  "Hohes Personenaufkommen": "Volume elevato di persone",
+  "Mittleres Personenaufkommen": "Volume medio di persone",
+  "Geringes Personenaufkommen": "Volume basso di persone",
+  "Sehr geringes Personenaufkommen": "Volume molto basso di persone",
   "Passagiere/Tag": "Passeggeri / giorno",
   "Damit sie ihre Zeichnung erneut bearbeiten können, müssen sie diesen Link speichern.": "Per modificare nuovamente il vostro disegno potete salvare questo link.",
   "Das Thema existiert nicht oder Sie haben nicht die notwendigen Rechte": "Il tema non esiste o l’utente non è provvisto dei diritti necessari",
