+++ src/layerInfos/BeleuchtungLayerInfo/lightingMapping.js
