+++ src/layerInfos/index.js
@@ -22,6 +22,7 @@ import RegionenkartePublicLayerInfo from './RegionenkartePublicLayerInfo';
 import NetzentwicklungTopicInfo from './NetzentwicklungTopicInfo';
 import NetzentwicklungLayerInfo from './NetzentwicklungLayerInfo';
 import BeleuchtungTopicInfo from './BeleuchtungTopicInfo';
+import BeleuchtungLayerInfo from './BeleuchtungLayerInfo';
 
 export { default as BetriebsRegionenLayerInfo } from './BetriebsRegionenLayerInfo';
 export { default as BuslinesLayerInfo } from './BuslinesLayerInfo';
@@ -47,6 +48,7 @@ export { default as RegionenkartePublicLayerInfo } from './RegionenkartePublicLa
 export { default as NetzentwicklungTopicInfo } from './NetzentwicklungTopicInfo';
 export { default as NetzentwicklungLayerInfo } from './NetzentwicklungLayerInfo';
 export { default as BeleuchtungTopicInfo } from './BeleuchtungTopicInfo';
+export { default as BeleuchtungLayerInfo } from './BeleuchtungLayerInfo';
 
 export default {
   BetriebsRegionenLayerInfo,
@@ -73,4 +75,5 @@ export default {
   NetzentwicklungTopicInfo,
   NetzentwicklungLayerInfo,
   BeleuchtungTopicInfo,
+  BeleuchtungLayerInfo,
 };
